+++ x/cronos/keeper/evmhooks_test.go
@@ -2,9 +2,13 @@ package keeper_test
 
 import (
 	"fmt"
+	"github.com/crypto-org-chain/cronos/app"
+	keepertest "github.com/crypto-org-chain/cronos/x/cronos/keeper/mock"
+	"github.com/crypto-org-chain/cronos/x/cronos/types"
 	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	cronosmodulekeeper "github.com/crypto-org-chain/cronos/x/cronos/keeper"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	gravitytypes "github.com/peggyjv/gravity-bridge/module/x/gravity/types"
@@ -30,7 +34,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				logs := []*ethtypes.Log{
 					{
 						Address: contract,
-						Topics:  []common.Hash{keeper.NativeTransferEvent.ID},
+						Topics:  []common.Hash{keeper.SendToAccountEvent.ID},
 					},
 				}
 				err := suite.app.EvmKeeper.PostTxProcessing(txHash, logs)
@@ -40,7 +44,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 		{
 			"not enough balance, expect fail",
 			func() {
-				data, err := keeper.NativeTransferEvent.Inputs.Pack(
+				data, err := keeper.SendToAccountEvent.Inputs.Pack(
 					recipient,
 					big.NewInt(100),
 				)
@@ -48,7 +52,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				logs := []*ethtypes.Log{
 					{
 						Address: contract,
-						Topics:  []common.Hash{keeper.NativeTransferEvent.ID},
+						Topics:  []common.Hash{keeper.SendToAccountEvent.ID},
 						Data:    data,
 					},
 				}
@@ -57,7 +61,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 			},
 		},
 		{
-			"success native transfer",
+			"success send to account",
 			func() {
 				suite.app.CronosKeeper.SetExternalContractForDenom(suite.ctx, denom, contract)
 				coin := sdk.NewCoin(denom, sdk.NewInt(100))
@@ -67,7 +71,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				balance := suite.app.BankKeeper.GetBalance(suite.ctx, sdk.AccAddress(contract.Bytes()), denom)
 				suite.Require().Equal(coin, balance)
 
-				data, err := keeper.NativeTransferEvent.Inputs.Pack(
+				data, err := keeper.SendToAccountEvent.Inputs.Pack(
 					recipient,
 					coin.Amount.BigInt(),
 				)
@@ -75,7 +79,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				logs := []*ethtypes.Log{
 					{
 						Address: contract,
-						Topics:  []common.Hash{keeper.NativeTransferEvent.ID},
+						Topics:  []common.Hash{keeper.SendToAccountEvent.ID},
 						Data:    data,
 					},
 				}
@@ -89,7 +93,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 			},
 		},
 		{
-			"failed ethereum transfer, invalid denom",
+			"failed send to ethereum, invalid gravity denom",
 			func() {
 				suite.SetupTest()
 
@@ -101,7 +105,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				balance := suite.app.BankKeeper.GetBalance(suite.ctx, sdk.AccAddress(contract.Bytes()), denom)
 				suite.Require().Equal(coin, balance)
 
-				data, err := keeper.EthereumTransferEvent.Inputs.Pack(
+				data, err := keeper.SendToEthereumEvent.Inputs.Pack(
 					recipient,
 					coin.Amount.BigInt(),
 					big.NewInt(0),
@@ -110,7 +114,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				logs := []*ethtypes.Log{
 					{
 						Address: contract,
-						Topics:  []common.Hash{keeper.EthereumTransferEvent.ID},
+						Topics:  []common.Hash{keeper.SendToEthereumEvent.ID},
 						Data:    data,
 					},
 				}
@@ -120,7 +124,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 			},
 		},
 		{
-			"success ethereum transfer",
+			"success send to ethereum",
 			func() {
 				suite.SetupTest()
 				denom := "gravity0x0000000000000000000000000000000000000000"
@@ -133,7 +137,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				balance := suite.app.BankKeeper.GetBalance(suite.ctx, sdk.AccAddress(contract.Bytes()), denom)
 				suite.Require().Equal(coin, balance)
 
-				data, err := keeper.EthereumTransferEvent.Inputs.Pack(
+				data, err := keeper.SendToEthereumEvent.Inputs.Pack(
 					recipient,
 					coin.Amount.BigInt(),
 					big.NewInt(0),
@@ -142,7 +146,7 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 				logs := []*ethtypes.Log{
 					{
 						Address: contract,
-						Topics:  []common.Hash{keeper.EthereumTransferEvent.ID},
+						Topics:  []common.Hash{keeper.SendToEthereumEvent.ID},
 						Data:    data,
 					},
 				}
@@ -157,6 +161,48 @@ func (suite *KeeperTestSuite) TestEvmHooks() {
 					SenderAddress: sdk.AccAddress(contract.Bytes()).String(),
 				})
 				suite.Require().Equal(1, len(rsp.SendToEthereums))
+			},
+		},
+		{
+			"failed send to ibc, invalid ibc denom",
+			func() {
+				suite.SetupTest()
+				// Create Cronos Keeper with mock transfer keeper
+				cronosKeeper := *cronosmodulekeeper.NewKeeper(
+					app.MakeEncodingConfig().Marshaler,
+					suite.app.GetKey(types.StoreKey),
+					suite.app.GetKey(types.MemStoreKey),
+					suite.app.GetSubspace(types.ModuleName),
+					suite.app.BankKeeper,
+					keepertest.IbcKeeperMock{},
+					suite.app.GravityKeeper,
+					suite.app.EvmKeeper,
+				)
+				suite.app.CronosKeeper = cronosKeeper
+
+				suite.app.CronosKeeper.SetExternalContractForDenom(suite.ctx, denom, contract)
+				coin := sdk.NewCoin(denom, sdk.NewInt(100))
+				err := suite.MintCoins(sdk.AccAddress(contract.Bytes()), sdk.NewCoins(coin))
+				suite.Require().NoError(err)
+
+				balance := suite.app.BankKeeper.GetBalance(suite.ctx, sdk.AccAddress(contract.Bytes()), denom)
+				suite.Require().Equal(coin, balance)
+
+				data, err := keeper.SendToIbcEvent.Inputs.Pack(
+					"recipient",
+					coin.Amount.BigInt(),
+				)
+				suite.Require().NoError(err)
+				logs := []*ethtypes.Log{
+					{
+						Address: contract,
+						Topics:  []common.Hash{keeper.SendToIbcEvent.ID},
+						Data:    data,
+					},
+				}
+				err = suite.app.EvmKeeper.PostTxProcessing(txHash, logs)
+				// should fail, because of not ibc denom name
+				suite.Require().Error(err)
 			},
 		},
 	}